requirements.txt
langchain
openai
supabase
chromadb
qdrant-client
python-telegram-bot
pydantic
python-dotenv 